[jenkins] Automated git mirroring from cr.imson.co repository

+++ cookbooks/volume-backup/files/pre-backup/backup-repo.sh
@@ -11,20 +11,20 @@ set -o errexit -o pipefail -o noclobber -o nounset
 
 # configuration
 . /srv/homelab/homelab_config.sh
-. $HOMELAB_CONFIG_PATH/backup_config.sh
+. "$HOMELAB_CONFIG_PATH/backup_config.sh"
 
 echo :: backing up homelab repo...
 
 # note: using -u because we need git to create the path, not mktemp...
-TEMP_REPO_PATH=$(mktemp -d -p $BACKUP_STAGING_PATH/)
+TEMP_REPO_PATH=$(mktemp -d -p "$BACKUP_STAGING_PATH/")
 
-pushd $TEMP_REPO_PATH/ > /dev/null
+pushd "$TEMP_REPO_PATH/" > /dev/null
 
-git clone --depth 1 $BACKUP_GIT_REPO .
+git clone --depth 1 "$BACKUP_GIT_REPO" .
 
 HOMELAB_REPO_SHA=$(git rev-parse --short HEAD)
-git archive --format=tar master | bzip2 -c > $BACKUP_STAGING_PATH/homelab-repo.$HOMELAB_REPO_SHA.tar.bz2
+git archive --format=tar master | bzip2 -c > "$BACKUP_STAGING_PATH/homelab-repo.$HOMELAB_REPO_SHA.tar.bz2"
 
 popd > /dev/null
 
-rm -rf $BACKUP_STAGING_PATH/$TEMP_REPO_PATH/
+rm -rf "${BACKUP_STAGING_PATH:?}/${TEMP_REPO_PATH:?}/"
